pyproject.toml
[project]
name = "m5-tdd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "bcrypt<4.0.0",
    "coverage>=7.10.7",
    "fastapi>=0.128.8",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.47",
    "uvicorn[standard]>=0.39.0",
]

[dependency-groups]
dev = [
    "faker>=37.12.0",
]
